requirements.txt
Flask
opencv-python
face_recognition
numpy
Pillow
python-dotenv
Flask-Cors
Flask-SQLAlchemy
pymysql
